RTree : force all node entries as Envelope

+++ modules/index/geotk-index-tree/src/main/java/org/geotoolkit/index/tree/hilbert/HilbertRTree.java
@@ -26,10 +26,13 @@ import org.geotoolkit.geometry.GeneralDirectPosition;
 import org.geotoolkit.geometry.GeneralEnvelope;
 import org.geotoolkit.index.tree.*;
 import org.geotoolkit.index.tree.calculator.Calculator;
+import org.geotoolkit.referencing.CRS;
 import org.geotoolkit.util.ArgumentChecks;
 import org.geotoolkit.util.collection.UnmodifiableArrayList;
 import org.geotoolkit.util.converter.Classes;
 import org.opengis.geometry.DirectPosition;
+import org.opengis.geometry.Envelope;
+import org.opengis.geometry.MismatchedReferenceSystemException;
 import org.opengis.referencing.crs.CoordinateReferenceSystem;
 import org.opengis.referencing.operation.TransformException;
 
@@ -76,7 +79,7 @@ public class HilbertRTree extends DefaultAbstractTree {
      * {@inheritDoc}
      */
     @Override
-    public void search(final GeneralEnvelope regionSearch, final List<GeneralEnvelope> result) throws TransformException {
+    public void search(final GeneralEnvelope regionSearch, final List<Envelope> result) throws TransformException {
         final CoordinateReferenceSystem entryCRS = regionSearch.getCoordinateReferenceSystem();
         if (!entryCRS.equals(crs)) {
             regionSearch.setEnvelope(Envelopes.transform(regionSearch, crs));
@@ -91,10 +94,9 @@ public class HilbertRTree extends DefaultAbstractTree {
      * {@inheritDoc}
      */
     @Override
-    public void insert(final GeneralEnvelope entry) throws TransformException {
-        final CoordinateReferenceSystem entryCRS = entry.getCoordinateReferenceSystem();
-        if (!entryCRS.equals(crs)) {
-            entry.setEnvelope(Envelopes.transform(entry, crs));
+    public void insert(final Envelope entry) throws MismatchedReferenceSystemException {
+        if(!CRS.equalsIgnoreMetadata(crs, entry.getCoordinateReferenceSystem())){
+            throw new MismatchedReferenceSystemException();
         }
         final DefaultNode root = getRoot();
         final int dim = entry.getDimension();
@@ -112,10 +114,9 @@ public class HilbertRTree extends DefaultAbstractTree {
      * {@inheritDoc}
      */
     @Override
-    public void delete(final GeneralEnvelope entry) throws TransformException {
-        final CoordinateReferenceSystem entryCRS = entry.getCoordinateReferenceSystem();
-        if (!entryCRS.equals(crs)) {
-            entry.setEnvelope(Envelopes.transform(entry, crs));
+    public void delete(final Envelope entry) throws MismatchedReferenceSystemException {
+        if(!CRS.equalsIgnoreMetadata(crs, entry.getCoordinateReferenceSystem())){
+            throw new MismatchedReferenceSystemException();
         }
         deleteHilbertNode(getRoot(), entry);
     }
@@ -127,7 +128,7 @@ public class HilbertRTree extends DefaultAbstractTree {
      * @param regionSearch area of search.
      * @param result {@code List} where is add search resulting.
      */
-    public static void searchHilbertNode(final DefaultNode candidate, final GeneralEnvelope regionSearch, final List<GeneralEnvelope> result) {
+    public static void searchHilbertNode(final DefaultNode candidate, final GeneralEnvelope regionSearch, final List<Envelope> result) {
         if (regionSearch.intersects(candidate.getBoundary(), true)) {
             if (candidate.isLeaf()) {
                 final List<DefaultNode> lN = (List<DefaultNode>) candidate.getUserProperty("cells");
@@ -157,7 +158,7 @@ public class HilbertRTree extends DefaultAbstractTree {
      * @param entry to insert.
      * @throws IllegalArgumentException if candidate or entry are null.
      */
-    public static void insertNode(final DefaultNode candidate, final GeneralEnvelope entry) {
+    public static void insertNode(final DefaultNode candidate, final Envelope entry) {
         ArgumentChecks.ensureNonNull("impossible to insert a null entry", entry);
         if (candidate.isFull()) {
             List<DefaultNode> lSp = splitNode(candidate);
@@ -177,12 +178,12 @@ public class HilbertRTree extends DefaultAbstractTree {
         }
         if (candidate.isLeaf()) {
             if ((!candidate.getBoundary().contains(entry, true))) {
-                List<GeneralEnvelope> lS = new ArrayList<GeneralEnvelope>();
+                List<Envelope> lS = new ArrayList<Envelope>();
                 searchHilbertNode(candidate, candidate.getBoundary(), lS);
                 lS.add(entry);
                 GeneralEnvelope envelope = DefaultTreeUtils.getEnveloppeMin(lS);
                 candidate.getTree().getCalculator().createBasicHL(candidate, (Integer) candidate.getUserProperty("hilbertOrder"), envelope);
-                for (GeneralEnvelope sh : lS) {
+                for (Envelope sh : lS) {
                     chooseSubtree(candidate, entry).getEntries().add(sh);
                 }
             } else {
@@ -210,13 +211,13 @@ public class HilbertRTree extends DefaultAbstractTree {
         }
         if (cleaf && (cHO < ((HilbertRTree) candidate.getTree()).getHilbertOrder())) {
 
-            final List<GeneralEnvelope> lS = new ArrayList<GeneralEnvelope>();
+            final List<Envelope> lS = new ArrayList<Envelope>();
             searchHilbertNode(candidate, candidate.getBoundary(), lS);
             if (lS.isEmpty()) {
                 throw new IllegalStateException("impossible to increase Hilbert order of a empty Node");
             }
             candidate.getTree().getCalculator().createBasicHL(candidate, cHO + 1, DefaultTreeUtils.getEnveloppeMin(lS));
-            for (GeneralEnvelope sh : lS) {
+            for (Envelope sh : lS) {
                 insertNode(candidate, sh);
             }
             return null;
@@ -489,7 +490,7 @@ public class HilbertRTree extends DefaultAbstractTree {
      * @throws IllegalArgumentException if entry is null.
      * @return subnode chosen.
      */
-    public static DefaultNode chooseSubtree(final DefaultNode candidate, final GeneralEnvelope entry) {
+    public static DefaultNode chooseSubtree(final DefaultNode candidate, final Envelope entry) {
         ArgumentChecks.ensureNonNull("impossible to choose subtree with entry null", entry);
 
         if (candidate.isLeaf() && candidate.isFull()) {
@@ -635,7 +636,7 @@ public class HilbertRTree extends DefaultAbstractTree {
      * @throws IllegalArgumentException if candidate or entry is null.
      * @return true if entry is find and deleted else false.
      */
-    private static void deleteHilbertNode(final DefaultNode candidate, final GeneralEnvelope entry) {
+    private static void deleteHilbertNode(final DefaultNode candidate, final Envelope entry) {
         ArgumentChecks.ensureNonNull("deleteHilbertNode Node2D candidate : ", candidate);
         ArgumentChecks.ensureNonNull("deleteHilbertNode Shape entry : ", entry);
         if (candidate.getBoundary().intersects(entry, true)) {
@@ -684,12 +685,12 @@ public class HilbertRTree extends DefaultAbstractTree {
 
             final HilbertRTree tree = (HilbertRTree) candidate.getTree();
             final Calculator calc = tree.getCalculator();
-            final List<GeneralEnvelope> lS = new ArrayList<GeneralEnvelope>();
+            final List<Envelope> lS = new ArrayList<Envelope>();
             searchHilbertNode(candidate, candidate.getBoundary(), lS);
             if (lS.size() <= tree.getMaxElements() * Math.pow(2, tree.getHilbertOrder() * 2) && !lS.isEmpty()) {
                 calc.createBasicHL(candidate, tree.getHilbertOrder(), DefaultTreeUtils.getEnveloppeMin(lS));
                 candidate.setUserProperty("isleaf", true);
-                for (GeneralEnvelope entry : lS) {
+                for (Envelope entry : lS) {
                     DefaultNode n = chooseSubtree(candidate, entry);
                     n.getEntries().add(entry);
                 }
@@ -712,7 +713,7 @@ public class HilbertRTree extends DefaultAbstractTree {
      * @return appropriate Node from tree.
      * @return
      */
-    public static DefaultNode createCell(final Tree tree, final DefaultNode parent, final DirectPosition centroid, final int hilbertValue, final List<GeneralEnvelope> entries, double... coordinates) {
+    public static DefaultNode createCell(final Tree tree, final DefaultNode parent, final DirectPosition centroid, final int hilbertValue, final List<Envelope> entries, double... coordinates) {
         final DefaultNode cell = new DefaultNode(tree, parent, centroid, centroid, null, entries);
         cell.setUserProperty("hilbertValue", hilbertValue);
         cell.setUserProperty("centroid", centroid);
@@ -723,7 +724,7 @@ public class HilbertRTree extends DefaultAbstractTree {
      * {@inheritDoc }
      */
     @Override
-    public DefaultNode createNode(Tree tree, DefaultNode parent, List<DefaultNode> listChildren, List<GeneralEnvelope> listEntries, double... coordinates) {
+    public DefaultNode createNode(Tree tree, DefaultNode parent, List<DefaultNode> listChildren, List<Envelope> listEntries, double... coordinates) {
         final int ddim = coordinates.length;
         if ((ddim % 2) != 0) {
             throw new IllegalArgumentException("coordinate dimension is not correct");
